Make two separate subclasses of SyncableBrowser
This allows both RawBrowserManager (QObject) and media browsers
(QWidgets) to cleanly inherit from SyncableBrowser. Also moved
BrowserSynchronizer to its own module.
